refactor: extract shared path utilities into path_utils module
Deduplicate is_alembic_migration() which was identically defined in
both dead_code.py and processes.py. Also move is_test_file() from
dead_code.py to the shared module since it was already imported
externally by mcp/tools.py with a private-name code smell.

Both functions now live in path_utils.py with public names, reducing
the risk of silent divergence if one copy is updated without the other.
